+++ gles/src/main/java/gamelauncher/gles/font/bitmap/GlyphData.java
@@ -5,8 +5,8 @@ public class GlyphData {
     public int width;
     public int height;
     public float advance;
-    public int bearingX;
-    public int bearingY;
+    public float bearingX;
+    public float bearingY;
 
     @Override public String toString() {
         return "GlyphData{" + "width=" + width + ", height=" + height + ", advance=" + advance + ", bearingX=" + bearingX + ", bearingY=" + bearingY + '}';
